Create README.md
E Notice Board App helps you access the online notices on your phone. This App is an online notice board maker where a group of people can easily communicate with each other through notices regarding education, technical and cultural events of an organisation. 
It is an android based application designed to run on smart phones. The notices can viewed by all members of the group.In this way the important notes can be broad-casted within less time.Notification can be sent out by notifying  the group members that a new notice has been posted  through alerts.
The interface of this application is straight forward and takes you roughly a minute to get started. Posting notices is easy, click on the post notice button and enter the text. Users can view the post on the spot by having a notification alert in android phone.Here registration is must for all the users having this application in order they want to have notifications.
Our android mobile app does the following:
•	Reduce wastage of time and energy.
•	Avoids duplication of data.
•	Searching for the notices is simple.
•	 Attention  to  each  and  every  notice  is  achieved  through  alerts.
•	It provides anytime and anywhere services.
•	It helps in keeping all the notices at one place.